Rename messagebus.js to messagebusclient.js
