+++ lib/messagebusclient.js
